Uruchamia się
