+++ src/main/java/pl/lublin/wsei/LoginScreenController.java
@@ -45,7 +45,7 @@ public class LoginScreenController {
     }
 
     public void loginButtonOnAction(){
-        FXMLLoader loader = new FXMLLoader(this.getClass().getResource("pl.lublin.wsei/menuScreen.fxml"));
+        FXMLLoader loader = new FXMLLoader(this.getClass().getResource("menuScreen.fxml"));
         Pane pane = getPane(loader);
         MenuScreenController menuScreenController = loader.getController();
         menuScreenController.setMainScreenController(mainScreenController);
