--- src/main/java/pl/lublin/wsei/Main.java
@@ -1,24 +1,0 @@
-package pl.lublin.wsei;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-
-
-public class Main extends Application {
-
-    @Override
-    public void start(Stage primaryStage) throws Exception{
-        Parent root = FXMLLoader.load(getClass().getResource("pl.lublin.wsei/mainScreen.fxml"));
-        primaryStage.setTitle("Kalkulator dietetyczny");
-        primaryStage.setScene(new Scene(root));
-        primaryStage.show();
-    }
-
-
-    public static void main(String[] args) {
-        launch(args);
-    }
-}
