+++ src/main/java/pl/lublin/wsei/PrimaryController.java
@@ -7,6 +7,6 @@ public class PrimaryController {
 
     @FXML
     private void switchToSecondary() throws IOException {
-        App.setRoot("secondary");
+       // App.setRoot("secondary");
     }
 }
