+++ src/main/java/pl/lublin/wsei/SecondaryController.java
@@ -7,6 +7,6 @@ public class SecondaryController {
 
     @FXML
     private void switchToPrimary() throws IOException {
-        App.setRoot("primary");
+        //App.setRoot("primary");
     }
 }
